3.6.0
git-svn-id: https://svn.code.sf.net/p/xsnow/code/trunk@439 79837d8a-40e5-47fd-869e-05d31bbd7fe9

+++ xsnow/src/wmctrl.h
@@ -23,7 +23,7 @@
 #include <X11/Xlib.h>
 #include "xsnow.h"
 
-extern int          GetCurrentWorkspace();
+extern long int     GetCurrentWorkspace();
 extern int          GetWindows(WinInfo **w, int *nw);
 extern Window       FindWindowWithName(Display *dsp, const char* needle);
 extern WinInfo     *FindWindow(WinInfo *windows, int nwin, Window id);
